project_v1/requirements.txt
Flask
numpy
pandas
matplotlib
seaborn
scipy
mlxtend
mpl_toolkits.mplot3d
networkx
scikit-learn
